# File: Remove `\r` in `get_as_text()` to keep standardized Unix format