This fixes a compatibility breakage from #63481.
See discussion in #63434.