@akien-mga 2022-07-31
> Also as discussed in the conversation, we want this change to apply only for `3.x` in order to avoid breaking compatibility.
> As I mentioned in the conversation, for `master` and onward, we should return the raw file as is.

Well I prefer fixing the compatibility breakage in `master` too because it was not an intentional design decision.

The `File` API for bindings was likely intentionally designed to provide users with Unix-formatted files, as evidenced by the explicit skip of `\r` in `get_line()`. This design decision can be challenged but that should be in its own proposal/PR, it shouldn't be done as a side effect of solving an Android-specific I/O performance regression.

For the same reason, I wouldn't change an API as core as `String::parse_utf8()` as part of this regression fix. First we fix the compat breakage (which is local to `File::get_as_text()`, and then we can discuss more changing to File and String APIs in follow-up proposals/PRs.

I don't think the cost of `replace("\n", "")` is too bad and in any case this should stay faster than the previous code, and functionally equivalent. So IMO that's good enough as a regression fix.