@akien-mga 2022-07-31
> I don't think the cost of `replace("\n", "")` is too bad and in any case this should stay faster than the previous code, and functionally equivalent. So IMO that's good enough as a regression fix.

I did a test on a very heavy file with a `debug` build (so not the most telling as not optimized), and it does seem to have a significant impact on such a pathological case:

```
$ file editor_translations.gen.h*
editor_translations.gen.h:     C source, ASCII text
editor_translations.gen.h_dos: C source, ASCII text, with CRLF line terminators

$ ls -lh editor_translations.gen.h*
-rw-r--r-- 1 akien akien 16M Jul 31 16:36 editor_translations.gen.h
-rw-r--r-- 1 akien akien 19M Jul 31 16:37 editor_translations.gen.h_dos

$ cat editor_translations.gen.h | wc -l
2952162
$ cat editor_translations.gen.h_dos | wc -l
2952162
```

Measuring conversion time with this script:
```gdscript
var f1 = File.new()
f1.open("editor_translations.gen.h", File.READ)

var start1 = Time.get_ticks_usec()
f1.get_as_text()
var end1 = Time.get_ticks_usec()

print("Unix file: %f s" % ((end1 - start1) / 1e6))


var f2 = File.new()
f2.open("editor_translations.gen.h_dos", File.READ)

var start2 = Time.get_ticks_usec()
f2.get_as_text()
var end2 = Time.get_ticks_usec()

print("DOS file:  %f s" % ((end2 - start2) / 1e6))
```

- Original `File::get_as_text()` from before #63481

```
Unix file: 3.107707 s
DOS file:  3.214315 s
```

- Current `master` after #63481

```
Unix file: 0.136303 s
DOS file:  0.162586 s
```

- This PR with `replace("\r", "")`

```
Unix file: 0.180824 s
DOS file:  1.624483 s
```

So it does make a significant difference on this extreme case. It's still much better than before #63481 especially when no replace is actually needed.

I'll give a try to implementing this in `parse_utf8` to compare.