@m4gr3d 2022-07-31
> But I don't know if we want to do this in something as low level as parse_utf8()

This should be fine if we put the logic behind an optional parameter which defaults to the current behavior.