@bruvzg 2022-07-31
> But I don't know if we want to do this in something as low level as parse_utf8()

If it's an optional parameter, why not, it should be fine.